[Refactor] Opt RuntimeFilter Phase 3
