+++ be/src/connector/connector.cpp
@@ -25,6 +25,7 @@
 #include "connector/jdbc_connector.h"
 #include "connector/lake_connector.h"
 #include "connector/mysql_connector.h"
+#include "exprs/runtime_filter_bank.h"
 
 namespace starrocks::connector {
