+++ be/src/exec/cross_join_node.h
@@ -17,6 +17,7 @@
 #include "column/vectorized_fwd.h"
 #include "exec/exec_node.h"
 #include "exprs/expr_context.h"
+#include "exprs/runtime_filter_bank.h"
 #include "gen_cpp/PlanNodes_types.h"
 #include "runtime/descriptors.h"
 #include "runtime/runtime_state.h"
