+++ be/src/formats/orc/orc_chunk_reader.h
@@ -21,7 +21,6 @@
 #include "common/object_pool.h"
 #include "exec/hdfs_scanner/hdfs_scanner.h"
 #include "exprs/expr.h"
-#include "exprs/runtime_filter_bank.h"
 #include "formats/orc/column_reader.h"
 #include "formats/orc/orc_mapping.h"
 #include "formats/orc/utils.h"
