+++ be/src/formats/orc/utils.h
@@ -20,7 +20,7 @@
 #include "cctz/civil_time.h"
 #include "cctz/time_zone.h"
 #include "column/vectorized_fwd.h"
-#include "exec/pipeline/operator.h"
+#include "exprs/runtime_filter_bank.h"
 #include "formats/orc/orc_mapping.h"
 #include "gen_cpp/orc_proto.pb.h"
 #include "io/shared_buffered_input_stream.h"
