+++ be/src/runtime/runtime_filter/runtime_filter_descriptor.h
@@ -1,0 +1,123 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <mutex>
+#include <unordered_set>
+#include <vector>
+
+#include "common/object_pool.h"
+#include "common/status.h"
+#include "exprs/expr_context.h"
+#include "gen_cpp/InternalService_types.h"
+#include "gen_cpp/RuntimeFilter_types.h"
+#include "gen_cpp/Types_types.h"
+#include "runtime/runtime_filter.h"
+#include "runtime/runtime_filter_layout.h"
+#include "types/logical_type.h"
+
+namespace starrocks {
+
+class HashJoinNode;
+class HashJoiner;
+class RuntimeState;
+
+// how to generate & publish this runtime filter
+// used in runtime filter build node. (TOPN/NLJoin/HashJoin)
+// in pipeline engine, all operators generated by the same factory use the same build descriptor.
+class RuntimeFilterBuildDescriptor : public WithLayoutMixin {
+public:
+    RuntimeFilterBuildDescriptor() = default;
+    Status init(ObjectPool* pool, const TRuntimeFilterDescription& desc, RuntimeState* state);
+    int32_t filter_id() const { return _filter_id; }
+    ExprContext* build_expr_ctx() { return _build_expr_ctx; }
+    LogicalType build_expr_type() const { return _build_expr_ctx->root()->type().type; }
+    int build_expr_order() const { return _build_expr_order; }
+    const TUniqueId& sender_finst_id() const { return _sender_finst_id; }
+    const std::unordered_set<TUniqueId>& broadcast_grf_senders() const { return _broadcast_grf_senders; }
+    const std::vector<TRuntimeFilterDestination>& broadcast_grf_destinations() const {
+        return _broadcast_grf_destinations;
+    }
+    bool has_remote_targets() const { return _has_remote_targets; }
+    bool has_consumer() const { return _has_consumer; }
+    const std::vector<TNetworkAddress>& merge_nodes() const { return _merge_nodes; }
+
+    TRuntimeFilterBuildType::type type() const { return _runtime_filter_type; }
+    bool is_asc() const { return _is_asc; }
+    bool is_nulls_first() const { return _is_nulls_first; }
+    size_t limit() const { return _limit; }
+
+    void set_runtime_filter(RuntimeFilter* rf) { _runtime_filter = rf; }
+    // used in TopN filter to intersect with other runtime filters.
+    void set_or_intersect_filter(RuntimeFilter* rf) {
+        std::lock_guard guard(_mutex);
+        if (_runtime_filter) {
+            _runtime_filter->intersect(rf);
+        } else {
+            _runtime_filter = rf;
+        }
+    }
+
+    // used in local group colocate runtime filter
+    void set_or_concat(RuntimeFilter* rf, int32_t driver_sequence) {
+        std::lock_guard guard(_mutex);
+        if (_runtime_filter == nullptr) {
+            _runtime_filter = rf;
+            _runtime_filter->group_colocate_filter().resize(_num_colocate_partition);
+        }
+        _runtime_filter->group_colocate_filter()[driver_sequence] = rf;
+    }
+
+    RuntimeFilter* runtime_filter() { return _runtime_filter; }
+    void set_is_pipeline(bool flag) { _is_pipeline = flag; }
+    bool is_pipeline() const { return _is_pipeline; }
+    // TRuntimeFilterBuildJoinMode
+    int8_t join_mode() const { return _join_mode; };
+    // only use when layout type == local colocate
+    size_t num_colocate_partition() const { return _num_colocate_partition; }
+    void set_num_colocate_partition(size_t num) { _num_colocate_partition = num; }
+    bool is_broad_cast_in_skew() const { return _is_broad_cast_in_skew; }
+
+private:
+    friend class HashJoinNode;
+    friend class HashJoiner;
+    int32_t _filter_id;
+
+    ExprContext* _build_expr_ctx = nullptr;
+    int _build_expr_order;
+    bool _has_remote_targets;
+    bool _has_consumer;
+    int8_t _join_mode;
+    TUniqueId _sender_finst_id;
+    std::unordered_set<TUniqueId> _broadcast_grf_senders;
+    std::vector<TRuntimeFilterDestination> _broadcast_grf_destinations;
+    std::vector<TNetworkAddress> _merge_nodes;
+    RuntimeFilter* _runtime_filter = nullptr;
+    bool _is_pipeline = false;
+    size_t _num_colocate_partition = 0;
+    // field used in top-n runtime filter
+    bool _is_asc{};
+    bool _is_nulls_first{};
+    size_t _limit{};
+
+    bool _is_broad_cast_in_skew = false;
+    int32_t _skew_shuffle_filter_id = -1;
+
+    TRuntimeFilterBuildType::type _runtime_filter_type;
+
+    std::mutex _mutex;
+};
+
+} // namespace starrocks
