+++ be/test/formats/parquet/file_reader_test.cpp
@@ -39,6 +39,7 @@
 #include "exprs/expr_executor.h"
 #include "exprs/expr_factory.h"
 #include "exprs/in_const_predicate.hpp"
+#include "exprs/runtime_filter_bank.h"
 #include "formats/parquet/column_chunk_reader.h"
 #include "formats/parquet/metadata.h"
 #include "formats/parquet/page_reader.h"
